Improve memory consumption
For every established GTP-U connection, several metric probes are
created upon connection. Among the created probes are 2 of type
"histogram". An histogram is represented as a process in the GTP-U-Node
which will be initialized with a `min_heap_size` of 40k WORDS (8byte on
64bit machines).

See [1] and [2] for a discussion about the issue.

Neither of the preconditions for increasing the `min_heap_size` are met:

* the histograms are kept over a long period of time (and are not short
  lived)
* the amount of processes keep growing with the amount of established
  connections

This commit upgrades the core dependency which set the `min_heap_size`
value to 233 words: `gtp_u_edp`. The other dependencies are upgraded
because they contain something similar (`eradius`) or because they have
almost a year of fixes ready to be used (`exometer_core`).

[1]: https://github.com/Feuerlabs/exometer_core/issues/100
[2]: http://erlang.org/doc/efficiency_guide/processes.html